requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-extensions==3.2.3
python-decouple==3.8
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-allauth==0.57.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.0
stripe==7.8.0
pandas==2.1.4
numpy==1.26.2
django-tenant-schemas==1.11.0
django-jsonfield==3.1.0
python-dateutil==2.8.2